% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noise.R
\name{Col_sign}
\alias{Col_sign}
\title{Calculates the p value of a timeseries according to the given colour}
\usage{
Col_sign(p, nu, reps)
}
\arguments{
\item{p}{double (data, including NAs if any)}

\item{nu}{numeric variable (0 for white, 1 for pink or 2 for brown)}

\item{reps}{numeric variable (number of replications)}
}
\description{
Calculates the p value of a timeseries according to the given colour
}
