% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noise.R
\name{noise.f}
\alias{noise.f}
\title{Creates a trajectory using coloured noise}
\usage{
noise.f(nu, tmax)
}
\arguments{
\item{nu}{numeric variable (type 0 for white, 1 for pink or 2 for brown)}

\item{tmax}{numeric variable (length of the trajectory)}
}
\description{
Creates a trajectory using coloured noise
}
